backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
torch==2.1.2
diffusers==0.25.0
transformers==4.36.2
Pillow==10.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
huggingface_hub==0.20.3
requests==2.31.0
